+++ lua/plugins/nvim-colorizer.lua
@@ -2,5 +2,16 @@
 return {
    "catgoose/nvim-colorizer.lua",
    event = { "BufReadPre", "BufNewFile" },
-   opts = {},
+   opts = {
+      options = {
+         parsers = {
+            css = true, -- preset: enables names, hex, rgb, hsl, oklch
+            tailwind = { enable = true, lsp = true, update_names = true },
+         },
+         -- display = {
+         --    mode = "virtualtext",
+         --    virtualtext = { position = "after" },
+         -- },
+      },
+   },
 }